requirements.txt
Django==1.11.13
django-bootstrap-datepicker-plus==3.0.4
django-bootstrap3==10.0.1
django-bootstrap4==0.0.6
mysqlclient==1.3.13
Pillow==5.1.0
pytz==2018.4
